Sua admin product

+++ app/Services/ProductService.php
@@ -4,35 +4,34 @@
 namespace App\Services;
 
 
-
 use App\Models\Product;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\File;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Pagination\Paginator;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\File;
 
 class ProductService
 {
 
     public function index()
     {
-        return Product::select('products.id', 'img', 'products.name as product_name', 'categories.name as category_name', 'price')
+        return Product::where('products.visible', FLAG_ON)->select('products.id', 'img', 'products.name as product_name', 'categories.name as category_name', 'price')
             ->join('categories', 'categories.id', '=', 'products.category_id')->paginate(PER_PAGE);
     }
 
     public function add($request)
     {
-
-        $data = $request->all();
-        $imgName = time() . $data['img']->getClientOriginalName();
-        $request->img->move(public_path('upload/product/'), $imgName);
+        $imgName = time() . $request['img']->getClientOriginalName();
+        $imgFile = $request['img'];
+        $imgFile->move(public_path('upload/product/'), $imgName);
         Product::create([
-            'name' => $data['name'],
-            'category_id' => $data['category_id'],
-            'description' => $data['description'],
-            'content' => $data['content'],
+            'name' => $request['name'],
+            'category_id' => $request['category_id'],
+            'description' => $request['description'],
+            'content' => $request['content'],
             'img' => $imgName,
-            'price' => $data['price'],
+            'price' => $request['price'],
+            'visible'=>FLAG_ON,
         ]);
     }
 
@@ -59,12 +58,21 @@ class ProductService
 
     public function edit($request, $product)
     {
+        $product = Product::find($product);
+
         $data = $request->all();
-        if (File::exists(public_path('upload/product/' . $product->img))) {
-            File::delete(public_path('upload/product/' . $product->img));
+
+        $imgName = '';
+        if (empty($data['img'])) {
+            $imgName = $product->img;
+        } else {
+            if (File::exists(public_path('upload/product/' . $product->img))) {
+                File::delete(public_path('upload/product/' . $product->img));
+            }
+            $imgName = time() . $data['img']->getClientOriginalName();
+            $request->img->move(public_path('upload/product/'), $imgName);
         }
-        $imgName = time() . $data['img']->getClientOriginalName();
-        $request->img->move(public_path('upload/product/'), $imgName);
+
         $product->update([
             'name' => $data['name'],
             'category_id' => $data['category_id'],
@@ -78,7 +86,8 @@ class ProductService
     public function delete($id)
     {
         $product = Product::find($id);
-        $product->delete();
+        $product->visible =FLAG_OFF;
+        $product->save();
     }
 
     public function hotProducts()
